extracted pre-processor for `$modes` replacement to its own
